### 2020-05-21 #2201 @whedon
[PRE REVIEW]: Automated Sleep Stage Scoring Using k-Nearest Neighbors Classifier
**Submitting author:** @teamPSG (<a href="http://orcid.org/0000-0001-6360-0714">Tamas Kiss</a>)
**Repository:** <a href="https://github.com/teamPSG/kNN_Sleep_Scorer_kNNSS" target ="_blank">https://github.com/teamPSG/kNN_Sleep_Scorer_kNNSS</a>
**Version:** v1.0.0
