### 2020-06-08 #2201 @teamPSG
Hi @oliviaguest! Previously (like 5+ years ago) I was working with Brian Harvey (brian.harvey@biogen.com) and Liam Scott (liamscott26.2@gmail.com) and we did automated sleep scoring together so they have the expertise to review this software. However, previously we published together so we are not entirely independent. I asked coauthors too, will be back with more names if I receive suggestions. Thanks for contacting me!
