### 2020-06-08 cbbc562
Adding important reference to JOSS paper
