+++ JOSS_Paper/paper.bib
@@ -176,3 +176,14 @@ author = "Stefano Bastianini and Chiara Berteotti and Alessandro Gabrielli and F
    Pages="35",
    doi={https://doi.org/10.3389/fphar.2018.00035}
 }
+
+@Article{barger2019,
+   Author="Barger, Z.  and Frye, C. G.  and Liu, D.  and Dan, Y.  and Bouchard, K. E. ",
+   Title="{{R}obust, automated sleep scoring by a compact neural network with distributional shift correction}",
+   Journal="PLoS ONE",
+   Year="2019",
+   Volume="14",
+   Number="12",
+   Pages="e0224642",
+   doi={https://doi.org/10.1371/journal.pone.0224642}
+}
