### 2020-06-08 #2201 @teamPSG
@oliviaguest I have received suggestions for potential reviewers from one of my coauthors. I hope one of them will take on the review. Please try these contacts: Jack Allocca, Ph.D., Univ Melbourne <giancarlo.allocca@gmail.com>; Giovanna Zoccoli, M.D., Ph.D., Univ Bologna <giovanna.zoccoli@unibo.it>; Richard Stephenson, Ph.D., Univ Toronto <richard.stephenson@utoronto.ca>; Steven Fox, Merck <steven_fox@merck.com>; Zeke Barger, Ph.D. <Zeke@berkeley.edu>. Hope this helps. Best, Tamás
