### 2020-06-25 #2377 @sbuergers
Hey @teamPSG,

This is a nicely written and well documented toolbox, well done! I used Matlab 2019b for testing and the [SoftwareVerification](https://github.com/teamPSG/kNN_Sleep_Scorer_kNNSS/tree/master/Software_Verification) ran without problems.

From what I could gather there are only a few points to consider:

**Contribution and authorship**: From the commit history it is unclear who, apart from Tamás Kiss, contributed to the software. Can you include a statement of contributions for the other authors?

**References**: The references in the main text do not render properly (e.g. [@STEPHENSON2009263,@BASTIANINI2014277...]). It might be that this can simply be fixed by putting a semicolon instead of a comma in-between authors, see [this guide](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax).

**Community guidelines**: I did not see clear guidelines for third-parties wishing to: i.) Contribute to the software, ii.) Report issues or problems with the software (other than contacting the author, but it would be useful to have some pointers for how to use github for this). There are plenty of examples here on github that do this well, for instance [here](https://github.com/TomKellyGenetics/graphsim).

@oliviaguest, thanks for setting us up! I take it these comments should not go in the issues of the project?

Cheers,
Steffen.
