### 2020-06-25 2430430
JOSS paper edit
Trying to fix incorrect rendering of references in JOSS paper.
