+++ JOSS_Paper/paper.bib
@@ -93,8 +93,8 @@ url = "http://www.sciencedirect.com/science/article/pii/S0165027009004476",
 author = "Richard Stephenson and Aimee M. Caron and Daniel B. Cassel and Jennifer C. Kostela"
 }
 
-@article{BASTIANINI2014277,
-title = "SCOPRISM: A new algorithm for automatic sleep scoring in mice",
+@Article{BASTIANINI2014277,
+title = "{SCOPRISM}: A new algorithm for automatic sleep scoring in mice",
 journal = "Journal of Neuroscience Methods",
 volume = "235",
 pages = "277--284",
