### 2020-06-26 1b66061
Edits related to JOSS submission
- Adding author contribution section to JOSS_Paper/paper.md
- Adding CONTRIBUTING.md to provide community guidelines
- Spell checking README.md
