+++ CONTRIBUTING.md
@@ -1,0 +1,15 @@
+# Contributing
+
+Thank you for helping to make this package better!
+
+If you use our code and run into a bug or any problem, please open an issue here. Any questions, comments or suggestions can be sent via GitHub or send an e-mail to Tamás Kiss at kiss.t (at) wigner.hu.
+
+Contributions to the code are very welcome. Since this repository follows the standard open source protocol, please use the basic [GitHub workflow](https://guides.github.com/introduction/flow/) to suggest a modification or addition. TL;DR: fork this repository and clone it to your own machine. Then make your changes, push your work back up to the fork on GitHub and open a [pull request](https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/about-pull-requests) so that we can review your changes and discuss them with you before accepting and merging them to the main project.
+
+Please restrict your commit to one modification per commit and ideally form a legible commit history. In your pull request, please describe clearly and concisely your motivation and the feature you developed.
+
+Also, please try to keep your code easily readable and tidy using [smart indenting](https://www.mathworks.com/help/matlab/matlab_prog/improve-code-readability.html). We feel that writing help to functions greatly improves understanding of the code and its function so please follow the MATLAB basic structure for `help` text: [Add Help for Your Program](https://www.mathworks.com/help/matlab/matlab_prog/add-help-for-your-program.html).
+
+## Code of conduct
+
+Please be polite :-) For a more detailed version, please see [Contributor Covenant Code of Conduct](https://www.contributor-covenant.org/version/2/0/code_of_conduct).
