### 2020-06-26 #2377 @teamPSG
Hi @sbuergers,

Thanks for your feedback and really fast action! Glad you like the toolbox and thanks for the suggestions.

**Contributions and authorship:** I included a section on who did what at the end of the manuscript.

**References:** thanks for the hint, it worked, references are now rendered properly (during writing I checked the ms offline using the [local compiler](https://gist.github.com/arfon/09a11047de1c972e37f77e95d2c10efa) and it looked good there -- some components of the compilation workflow might have changed in the course of development...)

**Community guidelines:** good point, I missed this one. I added a [CONTRIBUTING.md](https://github.com/teamPSG/kNN_Sleep_Scorer_kNNSS/blob/master/CONTRIBUTING.md) to the repo.

I guess the comments are fine right here. Thanks for the review!

Best,
 Tamás
