### 2020-07-06 #2377 @trisbek
I have not been able to run anything on my house PC so I will check with
Samika later (she is in californian time) and let you know/
 tris

El lun., 6 jul. 2020 a las 9:42, teamPSG (<notifications@github.com>)
escribió:

> Hi @oliviaguest <https://github.com/oliviaguest>, have you received
> feedback on ETA from @Emma-k-ward <https://github.com/Emma-k-ward> and/or
> @trisbek <https://github.com/trisbek>? (The text editor doesn't
> auto-complete the mention for_trisbek_ for me -- does it mean anything?)
> Thanks!
>
