### 2020-07-14 #2377 @trisbek
I have added comments but was unable to tick the boxes on read the PDF today

El lun., 13 jul. 2020 a las 11:50, Olivia Guest (<notifications@github.com>)
escribió:

> OK, no problem and let us wait for the reviewers to get back to us. If
> they don't give a vague ETA within a week or so, I'll email them. ☺️
>
