### 2020-07-14 #2377 @trisbek
Invite me again?
[image: image.png]

El mar., 14 jul. 2020 a las 14:58, whedon (<notifications@github.com>)
escribió:

> The reviewer already has a pending invite.
>
> @trisbek <https://github.com/trisbek> please accept the invite by
> clicking this link:
> https://github.com/openjournals/joss-reviews/invitations
>
