### 2020-07-14 #2377 @oliviaguest
Perhaps another @openjournals/joss-editors can lend me a helping hand, please? I assume that @whedon is correct and all @trisbek needs to do is click that link? ☺️
