### 2020-07-14 #2377 @danielskatz
right - clicking the link should work
