### 2020-07-14 #2377 @trisbek
Hi,

it does not let me accept and it does not let me see the pdf
[image: image.png]

Samika said, next time she will upload her scripts and OSF and the lot to a
new account of github but this time she is happy for me to copypaste her
comments.

El mar., 14 jul. 2020 a las 15:57, Olivia Guest (<notifications@github.com>)
escribió:

> Also @trisbek <https://github.com/trisbek> (sorry I keep asking!) can you
> clarify if Samika wishes to be included as a reviewer? How much of this
> code and paper have they reviewed, and do they want credit attributed to
> them for their work?
>
