### 2020-07-15 #2377 @oliviaguest
In case people are reading this in the future for inspiration: please be aware that what happened here is both highly anomalous and frowned upon. JOSS expects reviewers to write their own reviews after, e.g., running the code on their own hardware, etc. Delegation of review to a third party, especially when credit is unallocated, is not inline with our principles and we will be soon updating our guidelines to make this as clear as possible. This is merely being underlined here to avoid setting a bad precedent.
