### 2020-07-15 #2377 @trisbek
A key decision and feature to take into account by PSG is to decide if this is good enough for classic EEG human data. For this package to be useful to the thousands of researchers working in EEG human sleep data it would have to be validated in human datasets, beyond the datasets in animals already shown in the pdf and examples. A good dataset is the publicly available https://sleeptight.isr.uc.pt/?page_id=48
and also the recently made available consortium on the subject  https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4835314/
Would the authors care to comment how their package links and complements to this wide open initiative? It could make a big difference.
 This is a extent of influence comment, the documentation is clear, the software runs well.
