### 2020-07-15 #2377 @trisbek
Sure, but it was also open, I told you I would ask Samika, she accepted and
I told you that she had written the bit I posted a couple of days ago.
After your message she came to github and did it herself (good!). It is
better this way as it assigns credit where it is due. Now we have the right
type of format for my review and a separate one Samika's, as we agreed a
few weeks ago.

Hopefully both reviews are useful to @teamPSG in terms of putting their
tools in context to the eeg sleep community.

El mié., 15 jul. 2020 a las 10:49, Olivia Guest (<notifications@github.com>)
escribió:

> In case people are reading this in the future for inspiration: please be
> aware that what happened here is both highly anomalous and frowned upon.
> JOSS expects reviewers to write their own reviews after, e.g., running the
> code on their own hardware, etc. Delegation of review to a third party,
> especially when credit is unallocated, is not inline with our principles
> and we will be soon updating our guidelines to make this as clear as
> possible. This is merely being underlined here to avoid setting a bad
> precedent.
>
