### 2020-07-15 #2377 @oliviaguest
OK, I really feel the need to add one last clarification for future readers, the events are:

@trisbek decided he would ask @samikakumar and did without checking the exact details with me and if I am OK with that. This happened in private mostly. I am, and was, though OK with that, but I want(ed) to do this properly (i.e., give her credit). Therefore, I asked quite a few times of Tristan to tell Samika to join this review with her GitHub account. For whatever reason she didn't. Finally, Tristan explained that he copy-pasted (some of?) her words in his review. So ultimately, I think it's both fair, open, and ethical that Samika takes credit where she did work to improve this paper and code. This is what I asked for from the start (both in private and in this issue thread), as soon as I found out that she had been involved in this process: to include Samika officially, openly, and transparently.

Hope this helps to give the full backstory to readers.

Also thank you for apologising, @trisbek. Please however, let us continue with the actual matter at hand: reviewing this JOSS submission. 😁
