### 2020-07-15 #2377 @Emma-k-ward
Hi Olivia, authors,

This is a nice toolbox that was easy to use and gave me easily interpretable output with minimal time investment, even as a totally naive user, which I think was the authors' goal. Some comments on my experience and how I would think it could be slightly improved:

**Code**
I also ran the Software_Verification without issues when running code inside the functions, although a very minor problem came up: when I wrote a main script to call the individual functions it was hard to figure out where to set my working directory. I would usually write my own main script when downloading a package, set the directory and add any paths, and then call functions there, and this was a bit harder here because the directories are called in reference to their values in setup and I wasn't sure how to organise my own script then.

Once I solved that, I could run all the code and everything behaved as expected. I liked that I get output including the structures with original sleep stage labels (I assume this means the manually coded labels for the training) and predicted labels at the end, and the figures are nice. As a naive user I could get interpretable output very quickly, which I think was the goal of the toolbox.

In the code itself I miss some comments to make it clearer what is happening where. The comments that are there are nice, but very high level so I'd like to understand at first glance a little more of what's being calculated. For example, in FirstStep, we get quickly to:
%% Calculate features
files = dir([par.DataDir filesep '*' par.DataExt]);
for file = files'
    filename = strrep(file.name, par.DataExt, '');
    fprintf('Processing %s.\n', filename);
    generate_statespace(par.TopDir, filename, 'Description', filename, ...
        'IsTrainingSet', true, ...
        'OutDir', par.IntDir, 'WriteTmpFile', [filename par.PostFn]);
end

and inside generate_statespace a few comments say "[see notes inside]", but I don't see much in the way of notes.

**Documentation**
In the documentation, I do see a statement of need and a summary, but I would have liked to know a little bit more.  I think the argument is made very clearly that this process should be automated, and I appreciated hearing how the scoring is done manually. However, I don't think the paper is that clear about the advantages of this particular algorithm.

For example, the paper explains that this algorithm is preferable to the previously available algorithms because the features used are human-interpretable but I would've liked one more sentence about this. I only know from the paper that the features are based on power in relevant frequency bands and Hjorth parameters but I don't know in what sense a feature based on those characteristics is more human-interpretable. Are the features literally human-readable? For a single animal would I then be able to dig into the output of generate_statespace and see, as a very simplified example, that a sudden decrease in alpha and increase in delta power are key markers of a certain stage? If so, then I would emphasise this more because I think it's very interesting, but if not I don't understand the type of feature interpretability the authors mean.
I also found the Wikipedia citation odd, as I was expecting information about the interpretability of these specific features and instead was led to an article arguing for interpretability in general. Not a problem per se but rather vague. This is rather an important point in the philosophy behind the toolbox and I was also very interested, so I was surprised it was quickly brushed aside.

Then the point that the current method is more valid with animals after drug treatment, or humans with atypical sleep signatures, mostly rests on the argument that the transitions between states are less predictable in these groups, and therefore the current method drops any prior assumptions about transition rates and timing. The authors even state that there's potentially more sleep stages in these atypical cases, which I wouldn't have known and is interesting, although the current method only allows for 3 stages (wakeful, REM, non-REM) so it's less relevant. I would've expected either a reference here (fourth sentence before the heading Brief Software Description) or another sentence saying why it is a reasonable to have such a (lack of) assumption about transition rates. Or, alternatively, that in the output there was a metric showing how often the classifier transitioned between labels so researchers could assess whether they find that reasonable. I think this very small adjustment might be useful to people who want to evaluate whether they will switch to this method.

Overall these are small adjustments and I think it's a cool toolbox!

Best wishes,

Emma
