### 2020-07-24 #2377 @teamPSG
Hey @oliviaguest, I have a quick technical question regarding editing the paper: I'd like to indicate that some authors have changed their affiliations since the work was done. Like my affiliation should be Pfizer but my current affiliation is Wigner RCP. Is there a specific way to do it or as a workaround shall I just add "Current affiliation" in the `name` tag in the authors list?
