### 2020-07-24 #2377 @oliviaguest
@teamPSG do you want to have both previous and current? Or just change to current?
