### 2020-07-24 3f268a6
Modifications for JOSS paper review
In response to comments and suggestions from reviewers of the JOSS paper
a number of modifications were done:

- New example data added to Example_Data folder and updated
Example_Data/Readme.md to describe new data

- Changed scripts in Software_Verification directory to include code for
processing new data as well. Updated Software_Verification/Readme.md to
reflect changes. Added new result figure and performance log files in
Software_Verification and renamed old results figure and log.

- Added short description of hjorth.m in Function_Library/Readme.md

-
[…]
