+++ JOSS_Paper/paper.md
@@ -42,7 +42,7 @@ affiliations:
  - name: Center for Neuroscience, SRI International, Menlo Park, CA, USA
    index: 3
 
- - name: Takeda Pharmaceuticals, Inc., Cambridge, MA, USA
+ - name: Current affiliation -- Takeda Pharmaceuticals, Inc., Cambridge, MA, USA
    index: 4
 
 date: 1 May 2020
@@ -106,18 +106,28 @@ often becomes a bottleneck and a potential source of
 subjectivity affecting research outcomes.
 
 In this paper, we present an automated approach intended to eliminate
-these potential issues. Building on features classically extracted
-from EEG and EMG data and machine learning-based classification of
-PSG, this approach is capable of staging sleep in multiple species
-under control and drug-treated conditions, facilitating the detection
-of treatment-induced changes or other manipulations (e.g.,
-genetic). Using human interpretable features calculated from EEG and
-EMG will be important to understand drug mechanisms, for prediction of
-treatment outcomes, and as biomarkers or even translational
-biomarkers. However, the initial application of our approach is for
+these potential issues. The initial application of our approach is for
 basic and discovery research in which experiments are conducted in
 large cohorts of rodents, with the expectation that results can be
-translated to higher-order mammals or humans.
+translated to higher-order mammals or even humans. Building on
+features classically extracted from EEG and EMG data and machine
+learning-based classification of PSG, this approach is capable of
+staging sleep in multiple species under control and drug-treated
+conditions, facilitating the detection of treatment-induced changes or
+other manipulations (e.g., genetic). Using human interpretable
+features calculated from EEG and EMG will be important to understand
+drug mechanisms, for prediction of treatment outcomes, and as
+biomarkers or even translational biomarkers. For example, one of the
+features used by the algorithm is the power in the theta frequency
+band (called `eeg_theta` in the code), which is the 4 Hz to 12 Hz
+range and it is known that an increase of theta activity together with
+low EMG activity (our relevant features are called `emg_high` and
+`emg_RMS`) are the hallmark of REM sleep (see the
+[figure](https://en.wikipedia.org/wiki/Rapid_eye_movement_sleep#/media/File:Normal_EEG_of_mouse.png)
+in @wiki:REM). However, theta power is also associated with other
+phenomena, like anxiety [@John2014], thus our `eeg_theta` feature,
+besides being used for sleep scoring can also be used as a biomarker
+of drug effect.
 
 Multiple software applications have been developed to address the
 problem of automated sleep stage scoring. In their comparative review,
@@ -129,20 +139,20 @@ combining a convolutional neural network-based architecture to produce
 domain invariant predictions integrated with a hidden Markov model to
 constrain state dynamics based upon known sleep physiology. While our
 method also builds on machine learning techniques, it is based on
-interpretable features [@wiki:XAI] and uses a simpler algorithm for
-classification -- which should make it an ideal choice for the broader
-community as well as for sleep experts who might not be too familiar
-with complex machine learning approaches. Furthermore, we chose not to
-constrain the number of identifiable sleep/wake states or the
-probability of transition from one state to another, as we and others
-have found that drug interventions and disease processes tend to
-change not only the amount of different sleep stages but their
-transition probabilities as well. Finally, our method is a supervised
-method that requires a training set. While this might seem to be a
-disadvantage over non-supervised methods, we have found that drug
-treatment or pathological conditions can result in sleep stages not
-observed in healthy controls. Thus, the algorithm must be trained to
-these new stages.
+interpretable features and uses a simpler algorithm for classification
[…]
