+++ JOSS_Paper/paper.bib
@@ -71,6 +71,14 @@ author = "Reza Boostani and Foroozan Karimzadeh and Mohammad Nami"
     note = "[Online; accessed 24-April-2020]"
 }
 
+@misc{ wiki:REM,
+    author = "{Wikipedia contributors}",
+    title = "Rapid eye movement sleep --- {Wikipedia}{,} The Free Encyclopedia",
+    year = "2020",
+    howpublished = "\url{https://en.wikipedia.org/w/index.php?title=Rapid_eye_movement_sleep&oldid=966961239}",
+    note = "[Online; accessed 16-July-2020]"
+  }
+
 @misc{kNNSS,
     author = {Tamás Kiss and Stephen Morairty and Michael Schwartz and Thomas S. Kilduff and Derek L. Buhl and Dmitri Volfson},
     title = {{\textit{k}}NNSS: A {M}atlab package for automated sleep scoring using the k-nearest neighbors algorithm},
@@ -186,4 +194,55 @@ author = "Stefano Bastianini and Chiara Berteotti and Alessandro Gabrielli and F
    Number="12",
    Pages="e0224642",
    doi={https://doi.org/10.1371/journal.pone.0224642}
+}
+
+@Article{revel2012,
+   Author="Revel, F. G. and Moreau, J. L. and Gainetdinov, R. R and
+                  Ferragud, A. and {Velázquez-Sánchez}, C and
+                  Sotnikova, T. D. and Morairty, S. R. and
+                  Harmeier, A. and {Groebke Zbinden}, K. and Norcross,
+                  R. D. and Bradaia, A. and Kilduff, T. S. and
+                  Biemans, B. and Pouzet, B. and Caron, M. G. and
+                  Canales, J. J. and Wallace, T. L. and Wettstein,
+                  J. G. and Hoener, M. C.",
+   Title="{Trace amine-associated receptor 1 partial agonism reveals novel paradigm for neuropsychiatric therapeutics}",
+   Journal="Biol Psychiatry",
+   Year="2012",
+   Volume="72",
+   Number="11",
+   Pages="934--942",
+   doi={https://doi.org/10.1016/j.biopsych.2012.05.014}
+}
+
+@Article{John2014,
+   Author="John, T.  and Kiss, T.  and Lever, C.  and Érdi, P. ",
+   Title="{{A}nxiolytic drugs and altered hippocampal theta rhythms: the quantitative systems pharmacological approach}",
+   Journal="Network",
+   Year="2014",
+   Volume="25",
+   Number="1-2",
+   Pages="20--37",
+   doi={https://doi.org/10.3109/0954898x.2013.880003}
+}
+
+@Article{Harvey2013,
+   Author="Harvey, B. D.  and Siok, C. J.  and Kiss, T.  and Volfson, D.  and Grimwood, S.  and Shaffer, C. L.  and Hajós, M. ",
+   Title="{{N}europhysiological signals as potential translatable biomarkers for modulation of metabotropic glutamate 5 receptors}",
+   Journal="Neuropharmacology",
+   Year="2013",
+   Volume="75",
+   Pages="19--30",
+   Month="Dec",
+   doi={https://doi.org/10.1016/j.neuropharm.2013.06.020}
+}
+
+@Article{Mooij2020,
+   Author="{de Mooij}, S. M. M.  and Blanken, T. F.  and Grasman, R. P. P. P.  and Ramautar, J. R.  and Van Someren, E. J. W.  and van der Maas, H. L. J. ",
+   Title="{{D}ynamics of sleep: {E}xploring critical transitions and early warning signals}",
+   Journal="Comput Methods Programs Biomed",
+   Year="2020",
[…]
