### 2020-07-24 #2377 @teamPSG
@oliviaguest I uploaded the revised version implementing the "Current affiliation" part the simple way so that you can see what I mean exactly. Pls. let me know if this should be done differently.
