### 2020-07-24 #2377 @oliviaguest
@teamPSG ah, thanks! I will check and get back to you. In the meantime (no pressure asking to get organised), are you working on what @Emma-k-ward has pending (the items not checked off yet)?
