### 2020-07-24 #2377 @teamPSG
Hi @oliviaguest and Reviewers,

Authors would like to thank all Reviewers for their insightful comments and suggestions. I included here a list of answers and modifications describing accompanying commit: https://github.com/teamPSG/kNN_Sleep_Scorer_kNNSS/commit/3f268a639a482fc635a3c0fff5324adcb6deff6e

@trisbek and @samikakumar:
Re: Functionality/Performance -- we included a second dataset recorded from rats in control condition. The data files and manual scores are uploaded to GitHub and Readme.md files in the Software_Verification and the Example_Data folders were updated to reflect this. Furthermore, the three example scripts in Software_Verification were slightly modified to allow for a comfortable processing of the newly added data. Similarly, the JOSS paper was updated and now includes a figure about prediction accuracy of the rat data. Data from the non-human primate (NHP) recordings have not been published yet so we would not like to disclose that data or results at this time.

Re: Script to find model maximizing overall accuracy -- Thank you for this idea, it might be a good addition to the package on GitHub at a later time. In fact, in practical situations the experimenter might not want to maximize an average prediction accuracy. Indeed, the reason for adding the "deflated model" was to increase goodness of REM prediction at the expense of decreasing Wake and NREM prediction accuracy. Since Wake and NREM are much more numerous than REM a slight decrease of W/NR prediction brings overall scores (like Mathew's correlation coefficient) down. In short, each real-world research has its very specific questions and narrowing down which model to use depends very much on this question. So yes, it will be a good idea to collect a number of scripts that look for optimal models from a number of point of views.

Re: Using the tool for human EEG -- the package was initially developed to help pre-clinical drug research by enabling the uniform
scoring of data from mice, rats and NHPs as the drug candidate advances through the discovery pipeline. While we recognize the importance of developing tools for the analysis of human EEG, the focus of our work was to enable robust translatable automated sleep analysis for animal studies.

@Emma-k-ward:
Re: Code | path -- sorry for the confusion with setting up the path. I typically like to code the way I did it here, where I have a main `par` structure and use its fields later on in the script. I am wondering if combining the content of the `setup.m` file with your
main script (in the fashion you like it) would help you get this better organized.

Re: Code | comments -- I tried my best and added more comments in the most important functions :bulb:

Re: Documentation | features -- this is a very helpful comment, thank you Emma! I wrote the whole thing with electrophysiologists -- particularly those engaged with sleep research -- in mind. But you're right, people with a different background might (I hope they will) read into the paper and have the same questions. I added a few sentences in the paper to be more precise about what the features mean. Indeed, as you say, the features are literally human-readable: if I was to plot the `EEG_theta`, the `EEG_delta` and the `EMG_high` features I would be able to roughly point out when the animal is awake, when he is in non-REM or REM sleep. In rats and mice, in wakefulness there is typically high muscle (EMG) activity and the power in the EEG is low, except for the theta (4-12Hz) band when the animal is actively exploring its environment. Contrary to this, during
sleep, EMG is low. During REM sleep there is a gigantic theta peak, while during non-REM sleep delta is high. The point clouds in this 3 dimensional space are not clearly separate so the need for a classifier. Furthermore, other states such as cataplexy, might require other but similar features to be well classified. I also removed the reference to XAI Wiki page from the paper since as you pointed out it was not appropriately placed .

Re: Documentation | transitions & states -- the three "basic" stages are the stages that are most likely present in rodents. In human (and even in non-human primates) more stages can be classified. Drug manipulations might evoke still further states (my paper on this will be submitted immediately when this present methods paper is published and can be cited; I will link it to GitHub when published). Similarly, genetic modifications might evoke different sleep stages, like cataplexy in narcolepsy. We did not want to go into all these details in the paper because we felt that it would lead way too far but I added a few sentences in the manuscript clarifying these points a little more. Anyways, an important point that you might have missed is that the method allows for the classification of any number of states! It only depends on the training set: if an exotic state (like cataplexy) is first visually identified, then manually scored, then it will be trained and the algorithm will predict those scores as well. I added a comment on this in the paper. Also, I added references to the fourth sentence before "Brief Software Description" section to clarify the transition / state question.

We hope these modifications address Reviewers' comments and suggestions, however, let us know if further clarifications are needed.

Best regards,
 Tamás
