### 2020-07-30 #2377 @trisbek
I am OK and happy with the responses. I hope the tool is discovered,
recognised and used by animal sleep studies and associated questions. I
hope that some of the code and criteria also matches human studies in the
future as it would be useful for translational purposes. best to all!

El lun., 27 jul. 2020 a las 15:51, Olivia Guest (<notifications@github.com>)
escribió:

> @Emma-k-ward <https://github.com/Emma-k-ward> if you get a moment to
> check and tick off the final few boxes that would be amazing. 😊
>
