### 2020-08-04 #2377 @teamPSG
@oliviaguest in the meanwhile: is there a solution for how to properly format "Current affiliation" of authors (pls. see comments from 11 days ago)?
