### 2020-08-13 #2377 @teamPSG
Thanks @oliviaguest then I'll go with the solution I have for current affiliation.

@Emma-k-ward please let me know if I can do anything further to help you proceed with the review process. Looking back at the comments I submitted my responses almost 3 weeks ago...
