### 2020-08-17 #2377 @teamPSG
Thank you very much @oliviaguest! I was actually planning to ask you tomorrow is there is any deadline or targeted maximal length of a review. Also, what happens if a reviewer becomes silent during a review (anything might happen these days :mask:) -- like, can you make an editorial decision on publishing | rejecting a paper if one the reviewers doesn't finalize their review for a long time? But OK, let's wait for @Emma-k-ward for one more week and then I'll ask you the above questions :innocent:
