### 2020-08-25 #2377 @teamPSG
Hello @oliviaguest! I'd return to my questions from last week: I would like to ask you to please consider making a decision about acceptance of our paper as editor (or as board of editors). I am asking this because I feel the review of our paper got stuck. A brief summary of where we are: I submitted the paper on **May 20** (https://joss.theoj.org/papers/5726bdd4b144467c8093e90df0169c25), the review began June 22, I finished replying to comments July 24 and by today three of our four reviewers fully endorsed the paper, we are still waiting for an update from @Emma-k-ward who also mostly accepted our work. I am particularly asking for an editorial decision on our paper because the comments and suggestions only required minor modifications and additions that we have all completed more than a month ago. Thank you very much for considering my request.
