### 2020-09-02 #2377 @teamPSG
Hello @oliviaguest! Will this Zenodo archive work for the JOSS publication: https://zenodo.org/record/4009178#.X09YJ4uxX0O or with DOI https://doi.org/10.5281/zenodo.4009178

Thanks,
 Tamás
