### 2020-09-04 d8aa91e
Fixing DOIs
Fixing dois in JOSS_Paper/paper.bib to get JOSS publication right.
