+++ JOSS_Paper/paper.bib
@@ -19,7 +19,7 @@ number = "2",
 pages = "111--122",
 year = "1999",
 issn = "0165-0270",
-doi = "https://doi.org/10.1016/S0165-0270(99)00027-8",
+doi = "10.1016/S0165-0270(99)00027-8",
 url = "http://www.sciencedirect.com/science/article/pii/S0165027099000278",
 author = "Claude Robert and Christian Guilpin and Aymé Limoge"
 }
@@ -32,7 +32,7 @@ number = "2",
 pages = "131--148",
 year = "2000",
 issn = "1087-0792",
-doi = "https://doi.org/10.1053/smrv.1999.0087",
+doi = "10.1053/smrv.1999.0087",
 url = "http://www.sciencedirect.com/science/article/pii/S1087079299900874",
 author = "Thomas Penzel and Regina Conradt"
 }
@@ -44,7 +44,7 @@ volume = "140",
 pages = "77--91",
 year = "2017",
 issn = "0169-2607",
-doi = "https://doi.org/10.1016/j.cmpb.2016.12.004",
+doi = "10.1016/j.cmpb.2016.12.004",
 url = "http://www.sciencedirect.com/science/article/pii/S0169260716308276",
 author = "Reza Boostani and Foroozan Karimzadeh and Mohammad Nami"
 }
@@ -96,7 +96,7 @@ number = "2",
 pages = "263--274",
 year = "2009",
 issn = "0165-0270",
-doi = "https://doi.org/10.1016/j.jneumeth.2009.08.014",
+doi = "10.1016/j.jneumeth.2009.08.014",
 url = "http://www.sciencedirect.com/science/article/pii/S0165027009004476",
 author = "Richard Stephenson and Aimee M. Caron and Daniel B. Cassel and Jennifer C. Kostela"
 }
@@ -108,7 +108,7 @@ volume = "235",
 pages = "277--284",
 year = "2014",
 issn = "0165-0270",
-doi = "https://doi.org/10.1016/j.jneumeth.2014.07.018",
+doi = "10.1016/j.jneumeth.2014.07.018",
 url = "http://www.sciencedirect.com/science/article/pii/S0165027014002696",
 author = "Stefano Bastianini and Chiara Berteotti and Alessandro Gabrielli and Flavia {Del Vecchio} and Roberto Amici and Chloe Alexandre and Thomas E. Scammell and Mary Gazea and Mayumi Kimura and Viviana {Lo Martire} and Alessandro Silvani and Giovanna Zoccoli"
 }
@@ -120,7 +120,7 @@ author = "Stefano Bastianini and Chiara Berteotti and Alessandro Gabrielli and F
    Year="2019",
    Volume="13",
    Pages="207",
-   doi={https://doi.org/10.3389/fnins.2019.00207}
+   doi={10.3389/fnins.2019.00207}
 }
 
 @Article{vladimir2020,
@@ -131,7 +131,7 @@ author = "Stefano Bastianini and Chiara Berteotti and Alessandro Gabrielli and F
    Volume="337",
    Pages="108668",
    Month="May",
-   doi={https://doi.org/10.1016/j.jneumeth.2020.108668}
+   doi={10.1016/j.jneumeth.2020.108668}
 }
 
 @Article{gunn2020,
@@ -141,7 +141,7 @@ author = "Stefano Bastianini and Chiara Berteotti and Alessandro Gabrielli and F
    Year="2020",
    Pages="e12991",
    Month="Feb",
-   doi={https://doi.org/10.1111/jsr.12991}
+   doi={10.1111/jsr.12991}
 }
 
 @Article{zhang2020,
[…]
