### 2020-09-04 #2377 @oliviaguest
Cool! Somebody from @openjournals/joss-eics will take over this final step — so just sit tight. 😊
