### 2020-09-04 #2377 @teamPSG
Hello @kthyng! Thanks for guiding us through the last steps :world_map:

Version number is v1.0 and I created a release with this tag here: https://github.com/teamPSG/kNN_Sleep_Scorer_kNNSS/releases/tag/v1.0
