### 2020-09-04 #2377 @kthyng
Congrats on your new publication @teamPSG! Thanks to editor @oliviaguest and reviewers @Emma-k-ward, @sbuergers, @trisbek, and @samikakumar. This process couldn't happen without your time and hard work.

(will leave issue open until the doi resolves)
